requirements.txt
groq>=0.18.0
pydantic>=2.8.0
python-dotenv>=1.0.1
pandas>=2.2.2
rapidfuzz>=3.9.6
pytest>=8.3.2